+++ packages/expo-battery/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Migrated Android codebase to use Expo Modules API. ([#23883](https://github.com/expo/expo/pull/23883) by [@lukmccall](https://github.com/lukmccall))
+
 ## 7.4.0 — 2023-08-02
 
 _This version does not introduce any user-facing changes._